backend/requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
asyncpg==0.30.0
bcrypt==4.0.1 # passlibで__about__を使用するためにバージョンを固定
cryptography==44.0.2
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.1 # SQL Alchemyで非同期操作を行うための依存関係
httpx==0.28.1
passlib==1.7.4
pydantic==2.11.3
pydantic-settings==2.9.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-xdist==3.6.1
python-jose==3.4.0
python-multipart==0.0.20
SQLAlchemy==2.0.40
tzdata==2025.2
uvicorn==0.34.2